requirements.txt
umodbus~=1.0.4
paho-mqtt~=1.6.1
prometheus-client
pandas~=1.4.2
mysql-connector-python~=8.0.28
matplotlib~=3.5.1
numpy
astropy~=5.0.4
suntime~=1.2.5
pytz~=2022.1
httpx~=0.22.0
certifi~=2021.10.8
requests~=2.27.1